Add support for A+ LTI Tool v1.3 grading
If the A+ front sends the parameter `lti_launch_id` when
loading the assignment, then Acos-aplus adds the parameter to
the assignment iframe content in a hidden HTML `<input>` element.
When the student submits a solution, the LTI launch id is also
sent back to the A+ front in the grading data/payload.

Part of the issue
https://github.com/apluslms/a-plus/issues/1154